MS66 Renaming index file for Layout
